sosig

+++ code/modules/food_and_drinks/food/snacks_meat.dm
@@ -284,12 +284,16 @@
 
 /obj/item/reagent_containers/food/snacks/sausage/attackby(obj/item/I, mob/user, params)
 	if(istype(I, /obj/item/kitchen/knife))
-		if(!do_after(user, 1 SECONDS, src)) {
+		if(isturf(loc)) //CHECK IF ITS ON A TABLE OR OTHER SURFACE FOR THE LOVE OF GOD
+			if(!do_after(user, 1 SECONDS, src)) {
+				return
+			}
+			new /obj/item/reagent_containers/food/snacks/sausage/american(get_turf(user))
+			to_chat(user, span_notice("You snip [src]."))
+			qdel(src)
+		else
+			to_chat(user, span_warning("You need to put [src] on a surface to snip it!"))
 			return
-		}
-		new /obj/item/reagent_containers/food/snacks/sausage/american(loc)
-		to_chat(user, span_notice("You snip [src]."))
-		qdel(src)
 	else
 		..()
